Initial implementation of memory gathering
The config collector can now record the machines total, free and used
RAM and swap. It parses the output of the command `free`. These values are
parsed directly from the "Mem" and "Swap" rows of the output. For now, we
request the values in bytes so that we can store them as integers. In the
future we may look to record the human readable values by adding the `-h` flag
to the command and changing the data type to str.

Signed-off-by: Nick Child <nick.child@ibm.com>