Add function for running command and returning regex captures
Previously, if you wanted to run a command and get the regex, you would have
to run the command and then do all the regex work in the calling fuunction.
This commit adds a new method to the collector class that accepts 2 strings
and returns a string option. The inputs are the command and the regex string.
There is expected to be a capture group in the regex string. If the command
runs and a capture is found then the first capture is returned as a string.

The use of the new method has been implemented to all current data collecting functions

Signed-off-by: Nick Child <nick.child@ibm.com>